fixed game crashing when spawning more than 6 entities

+++ src/app.cpp
@@ -28,10 +28,10 @@ void App::Update()
     {
     case SCENE_TITLE:
         game->AddToTexlist( 10, 0, resources.title.id,
-                { windowWidth/2 - resources.title.width/2, 64},
-                { windowWidth/2 + resources.title.width/2, 64+resources.title.height},
-                {0,0},                           
-                {1,1});
+                { (float)windowWidth/2.f - resources.title.width/2.f, (float)64},
+                { (float)windowWidth/2.f + resources.title.width/2.f, (float)64+resources.title.height},
+                {0.f,0.f},                           
+                {1.f,1.f});
         if (ui->Button(*imdrawlist, resources.newGame, {windowWidth/2 - 100, windowHeight/2 - 50}, 200, 100, {1,1,1,0.5}) ||
             ImGui::IsKeyPressed(ImGuiKey_Enter, false))
             scene = SCENE_GAME;
